Roughly centre the page elements
This is simpler visually, as the story just runs from top to bottom,
rather than vertically and horizontally.

Sorry this is a big commit that also adapts the text sizes... but it
feels like part of this change.

Also, I've added styles to the main list item to the bottom of the document
that doesn't confirm to the style system of this project... I apologise.